Add configurable "max level" of headings to include
At the moment, we include all the headings that we can find in the
document when generating a table of contents. This is fine, but there
are some scenarios where we want to make this configurable – only render
headings up to h3s, for example, to reduce the size of the table of
contents.

This allows the user to pass an options hash to
`Helpers#table_of_contents`, which can include a parameter called
`max_level`. Any headings that are above this level will be ignored.